+++ GadgetEngine/Graphics/Text/TextMesh.cpp
@@ -23,23 +23,23 @@ TextMesh::~TextMesh(){
 	App::GetResourceManager().UnloadResource(fontName);
 }
 
-size_t TextMesh::GetTotalWidthInPixels() const{
-	size_t totalWidthInPixels = 0;
+int64_t TextMesh::GetTotalWidthInPixels() const{
+	int64_t totalWidthInPixels = 0;
 	for(const char c : text){
 		const FreetypeFontCharacter& ch = font->GetCharacters().at(c);
-		totalWidthInPixels += (static_cast<size_t>(ch.advanceX) >> 6);
+		totalWidthInPixels += ch.advanceX >> 6;
 	}
 
 	return totalWidthInPixels;
 }
 
-size_t TextMesh::GetTotalHeightInPixels() const{
-	size_t totalHeightInPixels = 0;
+int64_t TextMesh::GetTotalHeightInPixels() const{
+	int64_t totalHeightInPixels = 0;
 	for(const char c : text){
 		const FreetypeFontCharacter& ch = font->GetCharacters().at(c);
 
 		if(ch.rows > totalHeightInPixels){
-			totalHeightInPixels = static_cast<size_t>(ch.rows);
+			totalHeightInPixels = ch.rows;
 		}
 	}
